+++ agency/invoke.hpp
@@ -5,11 +5,18 @@
 #include <agency/detail/control_structures/bind.hpp>
 #include <agency/execution/executor/executor_traits.hpp>
 #include <agency/execution/executor/customization_points/sync_execute.hpp>
+#include <agency/detail/multi_function.hpp>
+#include <agency/detail/static_const.hpp>
 #include <agency/detail/type_traits.hpp>
 #include <utility>
 
 namespace agency
 {
+
+
+// default_invoke takes as parameters either
+// 1. an executor, a function, and an argument list or
+// 2. a function, and an argument list
 
 
 template<class Executor, class Function, class... Args,
@@ -19,7 +26,7 @@ __AGENCY_ANNOTATION
 detail::result_of_t<
   typename std::decay<Function&&>::type(typename std::decay<Args&&>::type...)
 >
-  invoke(Executor& exec, Function&& f, Args&&... args)
+  default_invoke(Executor& exec, Function&& f, Args&&... args)
 {
   auto g = detail::bind(std::forward<Function>(f), std::forward<Args>(args)...);
 
@@ -27,16 +34,81 @@ detail::result_of_t<
 }
 
 
+__agency_exec_check_disable__
 template<class Function, class... Args,
          __AGENCY_REQUIRES(!is_executor<typename std::decay<Function>::type>::value)
         >
+__AGENCY_ANNOTATION
 detail::result_of_t<
   typename std::decay<Function&&>::type(typename std::decay<Args&&>::type...)
 >
-  invoke(Function&& f, Args&&... args)
+  default_invoke(Function&& f, Args&&... args)
 {
   return std::forward<Function>(f)(std::forward<Args>(args)...);
 }
+
+
+// the agency::invoke customization point is defined below
+namespace detail
+{
+
+
+struct call_member_function_invoke
+{
+  __agency_exec_check_disable__
+  template<class Customizer, class... Args>
+  __AGENCY_ANNOTATION
+  constexpr auto operator()(Customizer&& customizer, Args&&... args) const ->
+    decltype(std::forward<Customizer>(customizer).invoke(std::forward<Args>(args)...))
+  {
+    return std::forward<Customizer>(customizer).invoke(std::forward<Args>(args)...);
+  }
+};
+
+
+struct call_free_function_invoke
+{
+  __agency_exec_check_disable__
+  template<class Customizer, class... Args>
+  __AGENCY_ANNOTATION
+  constexpr auto operator()(Customizer&& customizer, Args&&... args) const ->
+    decltype(invoke(std::forward<Customizer>(customizer), std::forward<Args>(args)...))
+  {
+    return invoke(std::forward<Customizer>(customizer), std::forward<Args>(args)...);
+  }
+};
+
+
+struct call_default_invoke
+{
+  template<class... Args>
+  __AGENCY_ANNOTATION
+  constexpr auto operator()(Args&&... args) const ->
+    decltype(agency::default_invoke(std::forward<Args>(args)...))
+  {
+    return agency::default_invoke(std::forward<Args>(args)...);
+  }
+};
+
+
+using invoke_t = multi_function<
+  detail::call_member_function_invoke,
+  detail::call_free_function_invoke,
+  detail::call_default_invoke
+>;
+
+
+} // end detail
+
+
+namespace
+{
+
+
+constexpr auto const& invoke = detail::static_const<agency::detail::invoke_t>::value;
+
+
+} // end anonymous namespace
 
 
 } // end agency